Create useActivityResources custom hook
